add dashboard page

+++ app/Http/Controllers/QuestionsController.php
@@ -5,8 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Gate;
 class QuestionsController extends Controller
 {
+
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -15,7 +20,8 @@ class QuestionsController extends Controller
     public function index()
     {
         // \DB::enableQueryLog();
-        $questions = Question::with('user')->latest()->paginate(10);
+        $questions = Question::with('user')->latest()->simplePaginate(10);
+        // dd($questions);
         return view('questions.index', compact('questions'));
         // dd(\DB::getQueryLog());
     }
@@ -40,7 +46,7 @@ class QuestionsController extends Controller
     public function store(AskQuestionRequest $request)
     {
         $request->user()->questions()->create($request->only('title', 'body'));
-        return redirect()->route('questions.index');
+        return redirect()->route('questions.index')->with('success', 'Your question has added');
     }
     /**
      * Display the specified resource.
@@ -62,7 +68,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize('update', $question);
         return view('questions.edit', compact('question'));
+        
+        
     }
 
     /**
@@ -74,8 +83,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize('update', $question);
         $question->update($request->only('title', 'body'));
-        return redirect('/questions')->with('success', "Your question has been updated");
+        return redirect('/questions')->with('success', 'Your question has been updated');
     }
 
     /**
@@ -86,7 +96,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize('delete', $question);
         $question->delete();
-        return redirect('/questions')->with('success', "Your question has been deleted.");
+        return redirect('/questions')->with('success', 'Your question has been deleted.');
     }
 }
